rename(route-folder): 공지 작성 컨벤션 반영
