+++ src/app/(admin)/admin/notice/write/_components/NoticeWriteForm/NoticeWriteForm.tsx
@@ -2,8 +2,9 @@
 
 import { UseFormReturn, useWatch } from "react-hook-form";
 import { NoticeWriteFormValues } from "../../_types/NoticeWriteType";
-import { TitleInput, ContentInput, CategoryInput, useSubmitNotice } from "./_internal";
+import { TitleInput, ContentInput, CategoryInput } from "./_internal";
 import { WriteImageSection, WriteActionSection } from "@/components";
+import useSubmitNotice from "../../_hooks/useSubmitNotice";
 
 const IMAGE_HELP_TEXT = "*사진은 최대 5장 첨부가 가능합니다. (선택)";
