+++ src/app/(admin)/admin/notice/write/_components/NoticeWriteForm/_internal/ContentInput/ContentInput.tsx
@@ -1,7 +1,7 @@
 import { cn } from "@/utils";
 import { RequiredText } from "@/components";
 import { useFormContext, useWatch } from "react-hook-form";
-import { NoticeWriteFormValues } from "../../../_types/NoticeWriteType";
+import { NoticeWriteFormValues } from "../../../../_types/NoticeWriteType";
 
 const ContentInput = () => {
   const { register, control } = useFormContext<NoticeWriteFormValues>();
