+++ src/app/(admin)/admin/notice/write/_components/NoticeWriteForm/_internal/index.ts
@@ -1,4 +1,3 @@
-export { default as TitleInput } from "./TitleInput";
-export { default as ContentInput } from "./ContentInput";
-export { default as CategoryInput } from "./CategoryInput";
-export { default as useSubmitNotice } from "./useSubmitNotice";
+export { default as TitleInput } from "./TitleInput/TitleInput";
+export { default as ContentInput } from "./ContentInput/ContentInput";
+export { default as CategoryInput } from "./CategoryInput/CategoryInput";
