+++ src/app/(admin)/admin/notice/write/_hooks/useSubmitNoticeEdit.ts
@@ -1,6 +1,6 @@
 import { usePutNoticeDetail } from "@/api/fetch/admin";
 import { useDeleteS3, usePostS3 } from "@/api/fetch/s3";
-import { NoticeEditFormValues } from "../../../_types/NoticeEditFormValues";
+import { NoticeEditFormValues } from "../[id]/_types/NoticeEditFormValues";
 
 interface UseSubmitNoticeEditParams {
   noticeId: number;
